requirements.txt
flask==3.0.0
google-auth-oauthlib==1.2.0
google-auth==2.27.0
google-api-python-client==2.116.0
